noop rename invite -> shareoptions
